% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_DMRn.R
\name{calculate_DMRn}
\alias{calculate_DMRn}
\title{Calculation of differential methylation values for DMRs}
\usage{
calculate_DMRn(
  DMRfile_list,
  beta,
  human_ref = c("Cancer", "Tissue"),
  Case_name,
  Control_name
)
}
\arguments{
\item{beta}{the beta matrix with case group and control group(The case group is in the front, the group group is in the back)}

\item{human_ref}{Human genome reference strand from different cancer tissue or different tissue}

\item{Case_name}{the colnames of case group}

\item{Control_name}{the colnames of control group}

\item{DMR}{the converted DMR that converted by function preinput}
}
\description{
This function calculates the differential methylation value of the DMR and the number of probes and CpGs included and the length of the DMR
}
\examples{
DMRfile_list = read.table(system.file("extdata","DMRfile.Array.list.txt",package = 'DMRmetric'))
beta_file = system.file("extdata","Array_beta.RDS",package = 'DMRmetric')
Array_beta = readRDS(beta_file)
for(i in 1:nrow(DMRfile_list)){DMRfile_list[i,1] = eval(parse(text=DMRfile_list[i,1]))}
case = colnames(Array_beta)[1:5]
control = colnames(Array_beta)[6:10]
metric.DMRn = calculate_DMRn(DMRfile_list = DMRfile_list,beta = Array_beta, human_ref = 'Tissue', Case_name=case, Control_name=control)

}
