% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_QnQl.R
\name{calculate_QnQl}
\alias{calculate_QnQl}
\title{calculate the metric Qn and Ql under different threshold for a DMR set.}
\usage{
calculate_QnQl(DMRfile_list, merged_file, method = c("1", "2"))
}
\arguments{
\item{DMRfile_list}{a file containing formated DMR files of different methods on the same dataset(for each DMR file, the format is : chr    start_pos    end_pos    ...).}

\item{merged_file}{a merged methylation file of group A and group B.}

\item{method}{1 or 2 (used in calculating the methylation difference, e.g. 1 ---count based; 2 ---methylation level based;   ) .}
}
\description{
This function will calculate the methyation difference and count the CpG number and the length in each DMR.
}
\examples{
DMRfile_list = read.table(system.file('extdata','DMRfile.BSseq.list.txt',package='DMRmetric'))
for(i in 1:nrow(DMRfile_list)){DMRfile_list[i,1] = eval(parse(text=DMRfile_list[i,1]))}
merged_example_file = readRDS(system.file('extdata','BSseq.merge.RDS',package='DMRmetric'))
metric.QnQl = calculate_QnQl(DMRfile_list,merged_example_file,method="1")

}
