% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_group_meth_inf.R
\name{merge_group_meth_inf}
\alias{merge_group_meth_inf}
\title{Merge the methylation report files of samples in each group into one methylation report file.}
\usage{
merge_group_meth_inf(list_methylation_report_file, meth, unmeth)
}
\arguments{
\item{list_methylation_report_file}{A file containing the methylation report files of samples with group id.}

\item{meth}{the column id to specify the meth counts.}

\item{unmeth}{the column id to specify the unmeth counts.}
}
\description{
It provides two ways of merging for each cytosine: 1. count based: merging the numbers of methylated and unmethylated cytosines from each sample; methylation level based: calculating the average methylation level based on the methylation levels from samples.
}
