% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_graph.R
\name{metric_graph}
\alias{metric_graph}
\title{Draw a line chart for DMR metric.}
\usage{
metric_graph(metric, type = c("Qn", "Ql", "DMRn"), Title = NULL)
}
\arguments{
\item{metric}{The text file containing the metric DMRn file path and the name of DMR detection method.}

\item{Title}{The title of rank graph, default NULL.}

\item{threshold}{The methylation difference threshold t, default 0-0.9.}
}
\description{
This function will draw a line chart for Qn, Ql and DMRn.
}
\examples{
library("ggplot2")
metric_graph(metric.DMRn,type="DMRn",Title='GroupA-GroupB')
}
